Revert "컴포넌트 정리"
This reverts commit 88cfe7a4c629bbface136d41791db67c76ae5836.
