+++ TcPCM_Connect/TcPCM_Connect.csproj
@@ -118,11 +118,11 @@
     <Reference Include="WindowsFormsIntegration" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Dashboard\DatailSearch.cs">
+    <Compile Include="Component\Select.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Dashboard\DatailSearch.Designer.cs">
-      <DependentUpon>DatailSearch.cs</DependentUpon>
+    <Compile Include="Component\Select.Designer.cs">
+      <DependentUpon>Select.cs</DependentUpon>
     </Compile>
     <Compile Include="Component\ConfigSetting.cs">
       <SubType>Form</SubType>
@@ -142,12 +142,6 @@
     <Compile Include="Controller\CalendarColumn.cs" />
     <Compile Include="Controller\CustomDataGridView.cs">
       <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="Controller\DateRangePanel.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="Controller\DateRangePanel.Designer.cs">
-      <DependentUpon>DateRangePanel.cs</DependentUpon>
     </Compile>
     <Compile Include="Controller\FlatCombo.cs">
       <SubType>Component</SubType>
@@ -283,8 +277,8 @@
     <Compile Include="xaml\UserControl2.xaml.cs">
       <DependentUpon>UserControl2.xaml</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="Dashboard\DatailSearch.resx">
-      <DependentUpon>DatailSearch.cs</DependentUpon>
+    <EmbeddedResource Include="Component\Select.resx">
+      <DependentUpon>Select.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Component\ConfigSetting.resx">
       <DependentUpon>ConfigSetting.cs</DependentUpon>
